FileManager - Update public API
