--- js/bundles/modules/parts/file_providers.js
@@ -1,11 +1,0 @@
-import DevExpress from './core';
-
-/// BUNDLER_PARTS
-/* fileProviders (dx.module-core.js) */
-
-import fileProviders from '../../../bundles/modules/file_providers';
-DevExpress.fileProviders = fileProviders;
-
-/// BUNDLER_PARTS_END
-
-module.exports = fileProviders;
