+++ js/file_management/custom_provider.js
@@ -1,10 +1,10 @@
-import { ensureDefined, noop } from '../../../core/utils/common';
-import { isFunction } from '../../../core/utils/type';
-import { compileGetter } from '../../../core/utils/data';
+import { ensureDefined, noop } from '../core/utils/common';
+import { isFunction } from '../core/utils/type';
+import { compileGetter } from '../core/utils/data';
 
-import { FileProvider } from './file_provider';
+import FileSystemProviderBase from './provider_base';
 
-class CustomFileProvider extends FileProvider {
+class CustomFileSystemProvider extends FileSystemProviderBase {
 
     constructor(options) {
         options = ensureDefined(options, { });
@@ -44,7 +44,7 @@ class CustomFileProvider extends FileProvider {
         return this._executeActionAsDeferred(() => this._renameItemFunction(item, name));
     }
 
-    createFolder(parentDir, name) {
+    createDirectory(parentDir, name) {
         return this._executeActionAsDeferred(() => this._createDirectoryFunction(parentDir, name));
     }
 
@@ -72,7 +72,7 @@ class CustomFileProvider extends FileProvider {
         return this._downloadItemsFunction(items);
     }
 
-    getItemContent(items) {
+    getItemsContent(items) {
         return this._executeActionAsDeferred(() => this._getItemsContentFunction(items));
     }
 
@@ -95,4 +95,4 @@ class CustomFileProvider extends FileProvider {
     }
 }
 
-module.exports = CustomFileProvider;
+module.exports = CustomFileSystemProvider;
