+++ js/file_management/utils.js
@@ -1,4 +1,4 @@
-import { each } from '../../core/utils/iterator';
+import { each } from '../core/utils/iterator';
 
 const PATH_SEPARATOR = '/';
 
@@ -73,23 +73,10 @@ const pathCombine = function() {
     return result;
 };
 
-const getDisplayFileSize = function(byteSize) {
-    const sizesTitles = [ 'B', 'KB', 'MB', 'GB', 'TB' ];
-    let index = 0;
-    let displaySize = byteSize;
-    while(displaySize >= 1024 && index <= sizesTitles.length - 1) {
-        displaySize /= 1024;
-        index++;
-    }
-    displaySize = Math.round(displaySize * 10) / 10;
-    return `${displaySize} ${sizesTitles[index]}`;
-};
-
 module.exports.getFileExtension = getFileExtension;
 module.exports.getName = getName;
 module.exports.getParentPath = getParentPath;
 module.exports.getPathParts = getPathParts;
 module.exports.getEscapedFileName = getEscapedFileName;
 module.exports.pathCombine = pathCombine;
-module.exports.getDisplayFileSize = getDisplayFileSize;
 module.exports.PATH_SEPARATOR = PATH_SEPARATOR;
