+++ js/ui/file_manager/ui.file_manager.js
@@ -45,7 +45,7 @@ class FileManager extends Widget {
         this._controller = new FileItemsController({
             currentPath: this.option('currentPath'),
             rootText: this.option('rootFolderName'),
-            fileProvider: this.option('fileProvider'),
+            fileProvider: this.option('fileSystemProvider'),
             allowedFileExtensions: this.option('allowedFileExtensions'),
             maxUploadFileSize: this.option('upload').maxFileSize,
             onSelectedDirectoryChanged: this._onSelectedDirectoryChanged.bind(this)
@@ -341,7 +341,7 @@ class FileManager extends Widget {
 
     _getDefaultOptions() {
         return extend(super._getDefaultOptions(), {
-            fileProvider: null,
+            fileSystemProvider: null,
 
             currentPath: '',
 
@@ -362,7 +362,7 @@ class FileManager extends Widget {
 
             toolbar: {
                 items: [
-                    'showNavPane', 'create', 'upload', 'viewSwitcher',
+                    'showNavPane', 'create', 'upload', 'switchView',
                     {
                         name: 'separator',
                         location: 'after'
@@ -457,11 +457,11 @@ class FileManager extends Widget {
                  */
                 move: false,
                 /**
-                 * @name dxFileManagerOptions.permissions.remove
+                 * @name dxFileManagerOptions.permissions.delete
                  * @type boolean
                  * @default false
                  */
-                remove: false,
+                delete: false,
                 /**
                  * @name dxFileManagerOptions.permissions.rename
                  * @type boolean
@@ -491,7 +491,7 @@ class FileManager extends Widget {
             case 'currentPath':
                 this._setCurrentPath(args.value);
                 break;
-            case 'fileProvider':
+            case 'fileSystemProvider':
             case 'selectionMode':
             case 'customizeThumbnail':
             case 'customizeDetailColumns':
