+++ testing/tests/DevExpress.ui.widgets/fileManager.tests.js
@@ -19,7 +19,6 @@ import './fileManagerParts/fileItemsController.tests.js';
 
 import './fileManagerParts/common.tests.js';
 
-import './fileManagerParts/ajaxProvider.tests.js';
 import './fileManagerParts/arrayProvider.tests.js';
 import './fileManagerParts/remoteProvider.tests.js';
 import './fileManagerParts/customProvider.tests.js';
