+++ testing/tests/DevExpress.ui.widgets/fileManagerParts/fileItemsController.tests.js
@@ -1,7 +1,7 @@
 const { test } = QUnit;
 
 import FileItemsController from 'ui/file_manager/file_items_controller';
-import { ErrorCode } from 'ui/file_manager/ui.file_manager.common';
+import ErrorCode from 'file_management/errors';
 import { createUploaderFiles, createUploadInfo, stubFileReader } from '../../../helpers/fileManagerHelpers.js';
 import { isString } from 'core/utils/type';
